pyproject.toml
[project]
name = "doc-support"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agentmail>=0.0.39",
    "flask>=3.1.1",
    "markdown>=3.8.2",
    "openai-agents>=0.1.0",
    "python-dotenv>=1.1.1",
]
